added real numbers (#16)
I decided not to use the Facebook API/SDK, mainly because it feels like to much overhead for one single number.
I do not want the struggles of data protection just to show off. We can manually update the number from time to time.

I added an endpoint for the artist count, though. Just for the lulz.